requirements.txt
gradio>=4.19.0
google-genai>=0.8.0
google-adk>=0.1.0
python-dotenv>=1.0.0
